fix(models): placing the interface in the sorrect file
