+++ angular-app/src/app/models/blog-info.ts
@@ -1,5 +1,3 @@
-import { Post } from './post';
-
 export interface BlogInfo {
   id: string;
   title: string;
@@ -23,12 +21,3 @@ export type BlogLinks = {
   mastodon: string;
 };
 
-export interface BlogPaginationInfo {
-  posts: Post[];
-  pagination: PageInfo;
-}
-
-export interface PageInfo {
-  hasNextPage: boolean;
-  endCursor: string;
-}
